+++ crates/uroborosql-fmt/src/visitor/expr/a_expr/like.rs
