+++ crates/uroborosql-fmt/src/visitor/expr/c_expr/func_expr/func_expr_common_subexpr/expr_as_typename_function.rs
