+++ crates/uroborosql-fmt/src/visitor/expr/c_expr/func_expr/func_expr_common_subexpr/only_keyword_function.rs
