+++ crates/uroborosql-fmt/src/visitor/expr/c_expr/func_expr/func_expr_common_subexpr/optional_iconst_function.rs
