+++ crates/uroborosql-fmt/src/visitor/expr/c_expr/func_expr/func_expr_common_subexpr/substring_function.rs
